+++ library/src/main/java/com/prolificinteractive/materialcalendarview/MonthView.java
@@ -1,7 +1,6 @@
 package com.prolificinteractive.materialcalendarview;
 
 import android.annotation.SuppressLint;
-import android.content.Context;
 import android.support.annotation.NonNull;
 import android.util.AttributeSet;
 import android.view.View;
@@ -9,13 +8,17 @@ import android.view.ViewGroup;
 import android.view.accessibility.AccessibilityEvent;
 import android.view.accessibility.AccessibilityNodeInfo;
 
+import com.prolificinteractive.materialcalendarview.MaterialCalendarView.ShowOtherDates;
 import com.prolificinteractive.materialcalendarview.format.DayFormatter;
 import com.prolificinteractive.materialcalendarview.format.WeekDayFormatter;
 
 import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.Collection;
 import java.util.List;
 
+import static com.prolificinteractive.materialcalendarview.MaterialCalendarView.SHOW_DEFAULTS;
+import static com.prolificinteractive.materialcalendarview.MaterialCalendarView.showOtherMonths;
 import static java.util.Calendar.DATE;
 import static java.util.Calendar.DAY_OF_WEEK;
 
@@ -30,12 +33,7 @@ class MonthView extends ViewGroup implements View.OnClickListener {
     protected static final int DEFAULT_MAX_WEEKS = 6;
     protected static final int DEFAULT_MONTH_TILE_HEIGHT = DEFAULT_MAX_WEEKS + 1;
 
-    public interface Callbacks {
-
-        void onDateChanged(CalendarDay date);
-    }
-
-    private Callbacks callbacks;
+    private MaterialCalendarView mcv;
 
     private final ArrayList<WeekDayView> weekDayViews = new ArrayList<>();
     private final ArrayList<DayView> monthDayViews = new ArrayList<>();
@@ -45,17 +43,17 @@ class MonthView extends ViewGroup implements View.OnClickListener {
 
     private final Calendar tempWorkingCalendar = CalendarUtils.getInstance();
 
-    private CalendarDay selection = null;
     private CalendarDay minDate = null;
     private CalendarDay maxDate = null;
 
-    private boolean showOtherDates = false;
+    private @ShowOtherDates int showOtherDates = SHOW_DEFAULTS;
 
     private final ArrayList<DecoratorResult> decoratorResults = new ArrayList<>();
 
 
-    public MonthView(Context context, CalendarDay month, int firstDayOfWeek) {
-        super(context);
+    public MonthView(@NonNull MaterialCalendarView view, CalendarDay month, int firstDayOfWeek) {
+        super(view.getContext());
+        this.mcv = view;
         this.month = month;
         this.firstDayOfWeek = firstDayOfWeek;
 
@@ -65,7 +63,7 @@ class MonthView extends ViewGroup implements View.OnClickListener {
         Calendar calendar = resetAndGetWorkingCalendar();
 
         for (int i = 0; i < DEFAULT_DAYS_IN_WEEK; i++) {
-            WeekDayView weekDayView = new WeekDayView(context, CalendarUtils.getDayOfWeek(calendar));
+            WeekDayView weekDayView = new WeekDayView(getContext(), CalendarUtils.getDayOfWeek(calendar));
             weekDayViews.add(weekDayView);
             addView(weekDayView);
             calendar.add(DATE, 1);
@@ -76,7 +74,7 @@ class MonthView extends ViewGroup implements View.OnClickListener {
         for(int r = 0; r < DEFAULT_MAX_WEEKS; r++) {
             for(int i = 0; i < DEFAULT_DAYS_IN_WEEK; i++) {
                 CalendarDay day = CalendarDay.from(calendar);
-                DayView dayView = new DayView(context, day);
+                DayView dayView = new DayView(getContext(), day);
                 dayView.setOnClickListener(this);
                 monthDayViews.add(dayView);
                 addView(dayView, new LayoutParams());
@@ -84,8 +82,6 @@ class MonthView extends ViewGroup implements View.OnClickListener {
                 calendar.add(DATE, 1);
             }
         }
-
-        setSelectedDate(CalendarDay.today());
     }
 
 
@@ -109,13 +105,16 @@ class MonthView extends ViewGroup implements View.OnClickListener {
         }
     }
 
-    public void setShowOtherDates(boolean show) {
-        this.showOtherDates = show;
+    public void setShowOtherDates(@ShowOtherDates int showFlags) {
+        this.showOtherDates = showFlags;
         updateUi();
     }
 
-    public boolean getShowOtherDates() {
-        return showOtherDates;
+    public void setSelectionEnabled(boolean selectionEnabled) {
+        for(DayView dayView : monthDayViews) {
+            dayView.setOnClickListener(selectionEnabled ? this : null);
+            dayView.setClickable(selectionEnabled);
+        }
     }
 
     public CalendarDay getMonth() {
@@ -134,7 +133,7 @@ class MonthView extends ViewGroup implements View.OnClickListener {
         int dow = CalendarUtils.getDayOfWeek(tempWorkingCalendar);
         int delta = firstDayOfWeek - dow;
         //If the delta is positive, we want to remove a week
-        boolean removeRow = showOtherDates ? delta >= 0 : delta > 0;
+        boolean removeRow = showOtherMonths(showOtherDates) ? delta >= 0 : delta > 0;
         if(removeRow) {
             delta -= DEFAULT_DAYS_IN_WEEK;
         }
@@ -184,9 +183,12 @@ class MonthView extends ViewGroup implements View.OnClickListener {
         updateUi();
     }
 
-    public void setSelectedDate(CalendarDay cal) {
-        selection = cal;
-        updateUi();
+    public void setSelectedDates(Collection<CalendarDay> dates) {
+        for(DayView dayView : monthDayViews) {
+            CalendarDay day = dayView.getDate();
+            dayView.setChecked(dates != null && dates.contains(day));
+        }
+        postInvalidate();
     }
 
     private void updateUi() {
@@ -194,7 +196,6 @@ class MonthView extends ViewGroup implements View.OnClickListener {
         for(DayView dayView : monthDayViews) {
             CalendarDay day = dayView.getDate();
             dayView.setupSelection(showOtherDates, day.isInRange(minDate, maxDate), day.getMonth() == ourMonth);
-            dayView.setChecked(day.equals(selection));
         }
         postInvalidate();
     }
@@ -212,28 +213,11 @@ class MonthView extends ViewGroup implements View.OnClickListener {
         }
     }
 
-    public void setCallbacks(Callbacks callbacks) {
-        this.callbacks = callbacks;
-    }
-
     @Override
     public void onClick(View v) {
         if(v instanceof DayView) {
-            for(DayView other : monthDayViews) {
-                other.setChecked(false);
-            }
             DayView dayView = (DayView) v;
-            dayView.setChecked(true);
-
-            CalendarDay date = dayView.getDate();
-            if(date.equals(selection)) {
-                return;
-            }
-            selection = date;
-
-            if(callbacks != null) {
-                callbacks.onDateChanged(dayView.getDate());
-            }
+            mcv.onDateClicked(dayView.getDate(), !dayView.isChecked());
         }
     }