+++ sample/src/main/java/com/prolificinteractive/materialcalendarview/sample/BasicActivity.java
@@ -5,11 +5,10 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.prolificinteractive.materialcalendarview.CalendarDay;
 import com.prolificinteractive.materialcalendarview.MaterialCalendarView;
-import com.prolificinteractive.materialcalendarview.OnDateChangedListener;
+import com.prolificinteractive.materialcalendarview.OnDateSelectedListener;
 import com.prolificinteractive.materialcalendarview.OnMonthChangedListener;
 
 import java.text.DateFormat;
@@ -21,7 +20,7 @@ import butterknife.ButterKnife;
 /**
  * Shows off the most basic usage
  */
-public class BasicActivity extends AppCompatActivity implements OnDateChangedListener, OnMonthChangedListener {
+public class BasicActivity extends AppCompatActivity implements OnDateSelectedListener, OnMonthChangedListener {
 
     private static final DateFormat FORMATTER = SimpleDateFormat.getDateInstance();
 
@@ -36,20 +35,27 @@ public class BasicActivity extends AppCompatActivity implements OnDateChangedLis
 
         widget.setOnDateChangedListener(this);
         widget.setOnMonthChangedListener(this);
+
+        //Setup initial text
+        textView.setText(getSelectedDatesString());
     }
 
     @Override
-    public void onDateChanged(@NonNull MaterialCalendarView widget, @Nullable CalendarDay date) {
-        if(date == null) {
-            textView.setText(null);
-        }
-        else {
-            textView.setText(FORMATTER.format(date.getDate()));
-        }
+    public void onDateSelected(@NonNull MaterialCalendarView widget, @Nullable CalendarDay date, boolean selected) {
+        textView.setText(getSelectedDatesString());
     }
 
     @Override
     public void onMonthChanged(MaterialCalendarView widget, CalendarDay date) {
-        Toast.makeText(this, FORMATTER.format(date.getDate()), Toast.LENGTH_SHORT).show();
+        //noinspection ConstantConditions
+        getSupportActionBar().setTitle(FORMATTER.format(date.getDate()));
+    }
+
+    private String getSelectedDatesString() {
+        CalendarDay date = widget.getSelectedDate();
+        if(date == null) {
+            return "No Selection";
+        }
+        return FORMATTER.format(date.getDate());
     }
 }