+++ sample/src/main/java/com/prolificinteractive/materialcalendarview/sample/CustomizeXmlActivity.java
@@ -1,39 +1,22 @@
 package com.prolificinteractive.materialcalendarview.sample;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
-import android.widget.TextView;
 
-import com.prolificinteractive.materialcalendarview.CalendarDay;
 import com.prolificinteractive.materialcalendarview.MaterialCalendarView;
-import com.prolificinteractive.materialcalendarview.OnDateChangedListener;
-
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
 
-public class CustomizeXmlActivity extends AppCompatActivity implements OnDateChangedListener {
-
-    private static final DateFormat FORMATTER = SimpleDateFormat.getDateInstance();
+public class CustomizeXmlActivity extends AppCompatActivity {
 
     @Bind(R.id.calendarView) MaterialCalendarView widget;
-    @Bind(R.id.textView) TextView textView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_customization);
         ButterKnife.bind(this);
-
-        widget.setOnDateChangedListener(this);
-    }
-
-    @Override
-    public void onDateChanged(@NonNull MaterialCalendarView widget, CalendarDay date) {
-        textView.setText(FORMATTER.format(date.getDate()));
     }
 
 }