+++ sample/src/main/java/com/prolificinteractive/materialcalendarview/sample/DynamicSettersActivity.java
@@ -1,24 +1,19 @@
 package com.prolificinteractive.materialcalendarview.sample;
 
-import android.app.AlertDialog;
 import android.app.DatePickerDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.widget.DatePicker;
 import android.widget.NumberPicker;
-import android.widget.Toast;
 
 import com.prolificinteractive.materialcalendarview.CalendarDay;
 import com.prolificinteractive.materialcalendarview.MaterialCalendarView;
-import com.prolificinteractive.materialcalendarview.OnDateChangedListener;
 
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Random;
 
@@ -27,9 +22,7 @@ import butterknife.ButterKnife;
 import butterknife.OnCheckedChanged;
 import butterknife.OnClick;
 
-public class DynamicSettersActivity extends AppCompatActivity implements OnDateChangedListener {
-
-    private static final DateFormat FORMATTER = SimpleDateFormat.getDateInstance();
+public class DynamicSettersActivity extends AppCompatActivity {
 
     @Bind(R.id.calendarView) MaterialCalendarView widget;
     private int currentTileSize;
@@ -41,17 +34,44 @@ public class DynamicSettersActivity extends AppCompatActivity implements OnDateC
         ButterKnife.bind(this);
 
         currentTileSize = MaterialCalendarView.DEFAULT_TILE_SIZE_DP;
-
-        widget.setOnDateChangedListener(this);
     }
 
-    @Override
-    public void onDateChanged(@NonNull MaterialCalendarView widget, @Nullable CalendarDay date) {
-        Toast.makeText(this, date == null ? "Unselected" : FORMATTER.format(date.getDate()), Toast.LENGTH_SHORT).show();
-    }
-
-    @OnCheckedChanged(R.id.check_other_dates) void onOtherMonthsChecked(boolean checked) {
-        widget.setShowOtherDates(checked);
+    @OnClick(R.id.button_other_dates) void onOtherDatesClicked() {
+        CharSequence[] items = {
+                "Other Months",
+                "Out Of Range",
+                "Decorated Disabled"
+        };
+        final int[] itemValues = {
+                MaterialCalendarView.SHOW_OTHER_MONTHS,
+                MaterialCalendarView.SHOW_OUT_OF_RANGE,
+                MaterialCalendarView.SHOW_DECORATED_DISABLED,
+        };
+        int showOtherDates = widget.getShowOtherDates();
+        boolean[] initSelections = {
+                MaterialCalendarView.showOtherMonths(showOtherDates),
+                MaterialCalendarView.showOutOfRange(showOtherDates),
+                MaterialCalendarView.showDecoratedDisabled(showOtherDates),
+        };
+        new AlertDialog.Builder(this)
+                .setTitle("Show Other Dates")
+                .setMultiChoiceItems(items, initSelections, new DialogInterface.OnMultiChoiceClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which, boolean isChecked) {
+                        int showOtherDates = widget.getShowOtherDates();
+                        if(isChecked) {
+                            //Set flag
+                            showOtherDates |= itemValues[which];
+                        }
+                        else {
+                            //Unset flag
+                            showOtherDates &= ~itemValues[which];
+                        }
+                        widget.setShowOtherDates(showOtherDates);
+                    }
+                })
+                .setPositiveButton(android.R.string.ok, null)
+                .show();
     }
 
     @OnCheckedChanged(R.id.check_text_appearance) void onTextAppearanceChecked(boolean checked) {
@@ -65,7 +85,7 @@ public class DynamicSettersActivity extends AppCompatActivity implements OnDateC
             widget.setDateTextAppearance(R.style.TextAppearance_MaterialCalendarWidget_Date);
             widget.setWeekDayTextAppearance(R.style.TextAppearance_MaterialCalendarWidget_WeekDay);
         }
-        widget.setShowOtherDates(checked);
+        widget.setShowOtherDates(checked ? MaterialCalendarView.SHOW_ALL : MaterialCalendarView.SHOW_NONE);
     }
 
     @OnClick(R.id.button_min_date) void onMinClicked() {
@@ -131,6 +151,24 @@ public class DynamicSettersActivity extends AppCompatActivity implements OnDateC
 
     @OnClick(R.id.button_clear_selection) void onClearSelection() {
         widget.clearSelection();
+    }
+
+    @OnClick(R.id.button_selection_mode) void onChangeSelectionMode() {
+        CharSequence[] items = {
+                "No Selection",
+                "Single Date",
+                "Multiple Dates"
+        };
+        new AlertDialog.Builder(this)
+                .setTitle("Selection Mode")
+                .setSingleChoiceItems(items, widget.getSelectionMode(), new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        widget.setSelectionMode(which);
+                        dialog.dismiss();
+                    }
+                })
+                .show();
     }
 
     private static final int[] DAYS_OF_WEEK = {